moved Sass files in to styles folder. Refactored Popup link by removing the button and adding the bind to the Link.
